refactor(Messages): Move messages into more specific subpackages
