+++ src/Server/Messages/MBJsonDeserializer.java
@@ -1,6 +1,11 @@
 package Server.Messages;
 
 import Game.Models.Map;
+import Server.Messages.REST.CreateLobby;
+import Server.Messages.REST.JoinLobby;
+import Server.Messages.REST.LobbyInfo;
+import Server.Messages.REST.ServerInfo;
+import Server.Messages.Socket.*;
 import com.google.gson.*;
 
 import java.lang.reflect.Type;
